refactor: rename tests/records.rs and extract smoke test to its own file
Splitting `test_smoke` into a separate `tests/smoke.rs` avoids multiple
logger
initializations within a single file, which is unsupported. This also
renames the
original file to `shadowing.rs` to better reflect its content.
